requirements.txt
streamlit
google-cloud-vision
python-docx
requests
python-dotenv
python-docx
google-generative-ai